requirements.txt
# requirements.txt
discord.py
aiohttp
python-dotenv
meteostat